FastApi/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
numpy==1.21.5
transformers==4.14.2
scipy==1.7.3
newspaper3k==0.2.8
nltk==3.6.5
